still struggling with axis
Former-commit-id: dba30324949b41a43fd223032dd7e276d7a7283a
Former-commit-id: e77296817622101e6ac7ce89d2cd433823d59992

+++ src/grpic/grpic_auxiliary.hpp
@@ -183,8 +183,8 @@ namespace ntt {
   // First calculation, with B0 and D
   template <>
   Inline void computeAuxH_D_B0<Dimension::TWO_D>::operator()(const index_t i, const index_t j) const {
-    real_t i_ {static_cast<real_t>(static_cast<long int>(i - N_GHOSTS))};
-    real_t j_ {static_cast<real_t>(static_cast<long int>(j - N_GHOSTS))};
+    real_t i_ {static_cast<real_t>(static_cast<int>(i) - N_GHOSTS)};
+    real_t j_ {static_cast<real_t>(static_cast<int>(j) - N_GHOSTS)};
 
     real_t h_11_ijP {m_mblock.metric.h_11({i_, j_ + HALF})};
     real_t h_22_iPj {m_mblock.metric.h_22({i_ + HALF, j_})};
